Major FHIM refactoring in order to harmonize FHIM and CIMI: Stage 3: Implement the ClinicalStatement/Topic/Context pattern
Added PatientEducationPerformedStmt class, which is now the new focal
class. Added PatientEducationPerformed class, which now holds the
Context information.  Renamed the PatientEducation class
PatientEducationTopic, which is now the Topic information. Made this
class a subtype of Intervention. Moved
PatientEducation.levelOfUnderstanding, location, professionalTime,
textProvided, and visit to PatientEducationPerformed because these are
Context properties.

Deleted PatientEducation.patient and .comment, because Patient and
CommentEvent are now accessible through the
PatientEducationPerformedStmt class Deleted
PatientEducation.counsellingRecipientCategory because this is now
accessible through the subjectOfRecord/subjectOfInformation construct in
the PatientEducationPerformedStmt class Deleted
PatientEducation.counsellingSubject because this is now accessible
through Intervention.Kind. Deleted PatientEducation.dateTimeEnd,
.dateTimeStart, .duration, and .performer because these are now
inherited from PerformedContext

Renamed PatientEducationTopic.counsellingBarrier to educationBarrier,
and .counsellingMethod to educationMethod because "education" is
preferred over "counselling" Renamed
PatientEducationTopic.educationalMaterialsProvided to
educationalMaterial to remove the context of "provision" from the Topic
Renamed PatientEducationPerformed.textProvided to communicationSummary
because the materials may not be written (e.g., multimedia) Renamed
PatientEducationPerformed.visit to encounter, because the education may
happen in inpatient situations (e.g., discharge instructions)