Changes to FHIM to harmonize to FHIR STU 3:
Added Schedule and Slot classes for FHIR
Added a Party class, which has an association to either Person or
Organization to accomodate a common pattern in FHIR. We may or may not
keep this.
Changed Network from being a subtype of organization to having an
association to Organization.  Added Network.networkIdentifier
Updated definitions