Refactoring to denormalize Topic/Context pattern
Changed a number of associations to ServiceDeliveryLocation
(InpatientMedicationDispense.ward, PharmacyDispense.dispenseToLocation,
PharmacyRequest.deliverTo, and PharmacyRequest.deliverToLocation) to
attributes in order to reduce diagram clutter.
Renamed OutpatientMedicationDispense.patientEducation to
.patientEducationPerformed and pointed it to PatientEducationPerformed
vice PatientEducationTopic due to PatientEducation refactoring
Made PharmacyRequest a subtype of Order
Made PharmacyPromise a subtype of Promise
Made PharmacyDispense and ReturnToStock subtypes of ActionPerformed
Renamed PharmacyDispense.id to ._id as this is likely a duplicate of
InformationEntry.identifier and will therefore be deleted soon.
Renamed PharmacyRequest.startDate to ._startDate and .stopDate to
._stopDate as these are likely duplicates of Order.dateTimeBegin and
.dateTimeEnd and will therefore be deleted soon.